Fejioa now builds static library + fixed warnings

+++ Fejioa/src/Platform/Windows/WindowsInput.cpp
@@ -28,7 +28,7 @@ namespace Fejioa
 		double xpos, ypos;
 		glfwGetCursorPos(window, &xpos, &ypos);
 
-		return std::make_pair(xpos, ypos);
+		return std::make_pair((float)xpos, (float)ypos);
 	}
 
 	float WindowsInput::GetMouseXImpl()
